+++ CHANGELOG.md
@@ -4,6 +4,23 @@
 
 ### Features
 
+- Swift Error Names (#2958)
+
+Instead of only the Swift error name and error code, the SDK now sends the Swift error name for enum-based errors.
+
+```Swift
+enum LoginError: Error {
+    case wrongUser
+    case wrongPassword
+}
+
+SentrySDK.capture(error: LoginError.wrongPassword)
+```
+
+Capturing the above Swift error will now result in the following error message in Sentry: `wrongPassword (Code: 1)` instead of only `(Code: 1)`.
+[Customized error descriptions](https://docs.sentry.io/platforms/apple/usage/#customizing-error-descriptions) have precedence over this feature.
+To avoid sending PII by accident, the SDK doesn't send the Swift error name for struct-based Swift errors, and the SDK drops the values of enums.
+
 - Create User and Breadcrumb from map (#2820)
 
 ### Fixes